Improves serial communication behaviour in case of error
In production mode, if connection could not be established, the system was endlessly looping trying again.

The connection is only established at MPF startup, should we later loose the serial port for some reason, the system would not
re-establish the connection by itself. Therefore, it is better to exit MPF should the connexion not be established, and manage a restart externally.

In case of a problem with a serial port, the warning message clearly indicates the faulty port.

In case of MPF shutdown due to comm failure, do not try to close the faulty connection which is already in the closing state (avoid cascading exception).

Regarding OPP platform:
- in the rare case where there is only one device, do not treat it as special case regarding serial
- if device does not report a serial number, default it to #0
- _poll_sender (opp.py):
    - optimize for speed
    - send pool before waiting response
- logs port name and serial number attached